packagescanner/requirements.txt
requirements-parser==0.2.0